chore(flex): code and proptype cleanup pass
