+++ packages/paste-core/utilities/flex/stories/index.stories.tsx
@@ -4,7 +4,8 @@ import {withKnobs, select, boolean, number, text} from '@storybook/addon-knobs';
 import {Box} from '@twilio-paste/box';
 import {Text} from '@twilio-paste/text';
 import {Paragraph} from '@twilio-paste/paragraph';
-import {Flex, Display, VerticalAlign, HorizontalAlign} from '../src';
+import {Flex} from '../src';
+import {Display, VerticalAlign, HorizontalAlign} from '../src/types';
 
 const flexDisplayOptions = ['flex', 'inline-flex'];
 const flexVerticalAlignOptions = ['top', 'center', 'bottom', 'stretch'];
